WIP: refactor of the events ui into smaller pieces
